test(structure): fix registration test after auth refactor and organize tests into logical packages
