+++ backend/src/test/java/com/justynagajdek/healthreservationsystem/integration/HealthReservationSystemApplicationTests.java
@@ -1,4 +1,4 @@
-package com.justynagajdek.healthreservationsystem;
+package com.justynagajdek.healthreservationsystem.integration;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
